feat: reveal custom path text field on error
Currently the UI expects the grpc server to only reply with an error
status when sending a reply with a custom path pattern, so the
corresponding error message is displayed as a validation string for the
text field.
However, it is possible that we receive errors for simpler replies as
well, and we should display those properly in the UI. As a workaround,
this PR simply ensures that the 'custom path' option gets selected and
the text field is revealed, if the server returns an error.