Fix desktop file Exec matching
snapd 2.72 removes the env BAMF=… prefix to all its binaries in the Exec
line. One consequence is that the desktop file becomes invalid. The Exec
is generated by snapd is pointing to an unexisting binary (as it’s a
systemd service and it doesn’t install in /snap/bin) and so, the file is
ignored.