fix: repack snapd snap to ensure latest/edge is always used
When the snapd .deb package has the same version number as the snapd
snap in latest/edge, then the .deb package will be used, even if
SNAPD_REEXEC=1 is set.

To fix this, download the snapd snap from latest/edge, unpack it,
set its major version to 1337, and repack it, then reinstall. This
ensures that the snap tooling will see that the snapd snap is newer than
the system .deb package and re-exec into the snap.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>