fix: use interface as deserialization tag for typed enums
For `TypedPromptFilter` and `TypedPrompt`, use the `"interface"` field
as the deserialization tag, and remove that field from the inner generic
types which get associated directly with the interface type.

It is nonsensical to have an interface field with a value which differs
from the interface type associated with its containing struct.

Importantly, this allows these typed structs to be deserialized to the
correct interface. Previously, each interface was tried in order, and
whichever succeeded first was used. This resulted in `TypedPromptFilter`
instances always being deserialized to the `Camera` variant, which
caused the filter to fail to match any prompt for another interface.

Notably, this means that the `"interface"` field is now required when
deserializing `TypedPromptFilter` and `TypedPrompt` values. It will
always be present for prompts, and it makes sense to require it for
prompt filters as well in order to parse the other fields correctly.

Lastly, add lots of tests related to filters deserialization and
matching.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>