fix: obey prompt filter during scripted client grace period
If a scripted client sequence ends with a grace period, continue to obey
the prompt filter when watching for any unexpected prompts.

Previously, the prompt filter was ignored when watching for prompts
during the grace period, meaning that any leftover prompt from any other
test would be caught during the grace period and cause the test to fail,
as well as interfere with that other test since the matching prompt
would be automatically denied.

The script from the prompt sequence must be exported to be used during
the grace period handler function.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>